+++ guillotina/contrib/dbusers/__init__.py
@@ -7,6 +7,7 @@ _ = MessageFactory("guillotina.contrib.dbusers")
 
 app_settings = {
     "auth_user_identifiers": ["guillotina.contrib.dbusers.users.DBUserIdentifier"],
+    "auth_validation_tasks": {"register_user": {"executor": "guillotina.contrib.dbusers.register_user"}},
 }