+++ stubs/jsonschema/exceptions/__init__.py
@@ -1,0 +1,7 @@
+class ValidationError(Exception):
+    
+    message: str
+    validator: str
+    validator_value: str
+    path: str
+    schema_path: str